+++ get.sh
@@ -7,8 +7,8 @@ export BINLOCATION="/usr/local/bin"
 version=$(curl -s https://api.github.com/repos/pixelandtonic/nitro/releases/latest | grep -i tag_name | sed 's/\(\"tag_name\": \"\(.*\)\",\)/\2/' | tr -d '[:space:]')
 
 if [ ! "$version" ]; then
-  echo "There was a problem trying to automatically install nitro. You can try to install manually:"
-  echo ""
+  echo "There was a problem trying to automatically install Nitro. You can try to install manually:"
+  echo
   echo "1. Open your web browser and go to https://github.com/pixelandtonic/nitro/releases"
   echo "2. Download the latest release for your platform and unzip it."
   echo "3. Run 'chmod +x ./nitro' on the unzipped 'nitro' executable."
@@ -20,6 +20,14 @@ hasCurl() {
   result=$(command -v curl)
   if [ "$?" = "1" ]; then
     echo "You need curl to use this script."
+    exit 1
+  fi
+}
+
+hasMultipass() {
+  result=$(command -v multipass)
+  if [ "$?" = "1" ]; then
+    echo "You need Multipass to use Nitro. Please install it for your platform https://multipass.run/"
     exit 1
   fi
 }
@@ -97,6 +105,7 @@ getNitro () {
   targetZipFile="$targetTempFolder"/$fileName
 
   echo "Downloading package $packageUrl to $targetZipFile"
+  echo
   curl -sSL "$packageUrl" --output "$targetZipFile"
 
   if [ "$?" = "0" ]; then
@@ -109,6 +118,7 @@ getNitro () {
 
     chmod +x ./nitro
     echo "Download complete."
+    echo
 
     if [ ! -w "$BINLOCATION" ]; then
       echo
@@ -119,10 +129,11 @@ getNitro () {
       echo "============================================================"
       echo
       echo "  sudo cp ./nitro $BINLOCATION/nitro"
+      echo "  nitro"
       echo
     else
       echo
-      echo "Running with sufficient permissions to attempt to move nitro to $BINLOCATION"
+      echo "Running with sufficient permissions to attempt to move the nitro executable to $BINLOCATION"
 
       if [ ! -w "$BINLOCATION/nitro" ] && [ -f "$BINLOCATION/nitro" ]; then
         echo
@@ -153,4 +164,5 @@ getNitro () {
 }
 
 hasCurl
+hasMultipass
 getNitro