+++ internal/cmd/root.go
@@ -11,7 +11,7 @@ import (
 var rootCmd = &cobra.Command{
 	Use:          "nitro",
 	Short:        "Local Craft CMS dev made easy",
-	Long:         `nitro is a command-line tool focused on making local Craft development quick and easy.`,
+	Long:         `Nitro is a command-line tool focused on making local Craft development quick and easy.`,
 	SilenceUsage: true,
 }
 
@@ -39,6 +39,7 @@ func init() {
 		hostsCommand,
 		contextCommand,
 		versionCommand,
+		selfUpdateCommand,
 	)
 	xdebugCommand.AddCommand(xdebugOnCommand, xdebugOffCommand, xdebugConfigureCommand)
 	machineCommand.AddCommand(destroyCommand, createCommand, restartCommand, startCommand, stopCommand)