+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes for Nitro
 
+## Unreleased
+
+### Added
+- Added a `self-update` command.
+- Add a Multipass check to get.sh.
+
 ## 0.7.5 - 2020-04-12
 
 ### Added